feat: split into modules all functions
