+++ src/archive/index.js
@@ -1,0 +1,2 @@
+export { compress } from "./compress.js"
+export { decompress } from "./decompress.js"
